requirements.txt
PyMuPDF>=1.24.0
google-genai>=1.0.0
tenacity>=8.2.0
python-dotenv>=1.0.0
